+++ app/javascript/components/lease_ups/lottery_results_page/utils/byFunction.js
