+++ app/javascript/packs/react_application.js
@@ -6,10 +6,10 @@ import ApplicationNewPage from 'components/applications/ApplicationNewPage'
 import ApplicationsPage from 'components/applications/ApplicationsPage'
 import FlaggedApplicationsIndexPage from 'components/applications/flagged/FlaggedApplicationsIndexPage'
 import FlaggedApplicationsShowPage from 'components/applications/flagged/FlaggedApplicationsShowPage'
+import LotteryResultsPdfGenerator from 'components/lease_ups/lottery_results_page/LotteryResultsPdfGenerator'
 import ListingApplicationsPage from 'components/listings/ListingApplicationsPage'
 import ListingPage from 'components/listings/ListingPage'
 import ListingsPage from 'components/listings/ListingsPage'
-import LotteryResultsPdfGenerator from 'components/LotteryResultsPdfGenerator/LotteryResultsPdfGenerator'
 import HomePage from 'components/pages/HomePage'
 import LeaseUpApp from 'routes/LeaseUpApp'
