+++ app/javascript/routes/LeaseUpRoutes.js
@@ -5,7 +5,7 @@ import { Routes, Route } from 'react-router-dom'
 import ApplicationPage from 'components/applications/ApplicationPage'
 import LeaseUpApplicationsPage from 'components/lease_ups/LeaseUpApplicationsPage'
 import LeaseUpListingsPage from 'components/lease_ups/LeaseUpListingsPage'
-import LotteryResultsPdfGenerator from 'components/LotteryResultsPdfGenerator/LotteryResultsPdfGenerator'
+import LotteryResultsPdfGenerator from 'components/lease_ups/lottery_results_page/LotteryResultsPdfGenerator'
 import SupplementalApplicationPage from 'components/supplemental_application/SupplementalApplicationPage'
 import appPaths from 'utils/appPaths'
